Added SafeStringBuilder class. (#55)
SafeStringBuilder class supports mixing tainted and non-tainted substring within one instance. For example, if a substring is added via to an instance of `SafeStringBuilder` via `Append(string value)` method, it will be considered tainted. On the other hand, a substring via `UncheckedAppend(string value)` method, it will be considered safe from being tampered by a potential attacker.
Other methods, like `Insert` or `Replace` also have an "unchecked" version (`UncheckedInsert` and `UncheckedReplace`).

Unlike in `SafeString.Format()` method, attack detection is postponed until a string is built inside an instance of `SafeStringBuilder` and `ToString()` method is called. If an attack is detected, an `AttackDetectedException` will be thrown.

Co-authored-by: Sergey Podkorytov <spodkorytov@ptsecurity.com>